Move high level function (ex. init_param_unicycle, VelToMotorReference, ...) in high_level_control.h. Change type message motor_control_t https://github.com/officinerobotiche/orblibcpp/issues/5 @Myzhar Build first version task controller in motorReference
