+++ src/control/motors_PID.c
@@ -13,7 +13,7 @@
  * WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
  * Public License for more details
-*/
+ */
 
 /******************************************************************************/
 /* Files to Include                                                           */
@@ -34,6 +34,7 @@
 #include <stdbool.h>         /* For true/false definition                     */
 #include <dsp.h>
 #include <pwm12.h>
+#include "control/high_level_control.h"
 #include "control/motors_PID.h"       /* variables/params used by motorsPID.c         */
 #include "system/system.h"
 #include "system/user.h"
@@ -59,51 +60,33 @@ fractional kCoeffs2[3]; //Coefficienti KP, KI, KD Per PID2 Destra
 int AdcBuffer[2][ADC_BUFF] __attribute__((space(dma), aligned(256)));
 
 /** */
-//From System
-extern parameter_system_t parameter_system;
 
 volatile int PulsEncL = 0; //Buffer for deadReckoning
 volatile int PulsEncR = 0; //Buffer for deadReckoning
 
 parameter_motor_t parameter_motor_left, parameter_motor_right;
-parameter_unicycle_t parameter_unicycle;
 constraint_t constraint;
-velocity_t vel_rif, vel_mis;
 pid_control_t pid_left, pid_right;
 enable_motor_t enable_motors;
 motor_t motor_left, motor_right;
 
-//variables for emergency
-emergency_t emergency;
-velocity_t last_vel_rif;
-bool save_velocity = true;
-
-float const_vel[3];
-
-//Definition value for parameter unicycle
-typedef struct parameter_unicycle_int {
-    long radius_l;
-    long radius_r;
-    long wheelbase;
-} parameter_unicycle_int_t;
 typedef struct parameter_vel {
     float k_vel;
     int sign;
 } parameter_vel_t;
-typedef struct motor_ref_int {
-    long int ref_con_left;
-    long int ref_con_right;
-} motor_ref_int_t;
 
 k_odo_t k_odo;
 float wheel_m;
 
-parameter_unicycle_int_t parameter_unicycle_int;
 parameter_vel_t motor_vel_left, motor_vel_right;
 
+/***/
+// New control variable
+#define NUM_MOTORS 2
+unsigned int control_motor_state[NUM_MOTORS];
+unsigned int control_state = 0;
 
 /**/
-
 extern volatile unsigned long timePeriodL; //Periodo Ruota Sinistra
 extern volatile unsigned long timePeriodR; //Periodo Ruota Destra
 extern volatile unsigned SIG_VELL; //Verso rotazione ruota Sinistra
@@ -115,10 +98,6 @@ extern volatile unsigned SIG_VELR; //Verso rotazione ruota Destra
 /******************************************************************************/
 
 void init_parameter(void) {
-    parameter_unicycle.radius_l = 0.04; //Radius left wheel
-    parameter_unicycle.radius_r = 0.04; //Radius right wheel
-    parameter_unicycle.wheelbase = 0.20; //Wheelbase
-    parameter_unicycle.sp_min = 0.0001; // FLT_MIN
     //Left motor parameters
     parameter_motor_left.k_vel = K_VEL; //Gain to convert input capture value to velocity
     parameter_motor_left.k_ang = K_ANG; //Gain to convert QEI value to rotation movement
@@ -130,14 +109,6 @@ void init_parameter(void) {
     parameter_motor_right.versus = false;
     parameter_motor_right.enable_set = false;
 
-    //parameter_motors.pwm_step = 4096;
-
-    update_parameter_unicycle();
-
-    vel_rif.v = 0;
-    vel_rif.w = 0;
-    vel_mis.v = 0;
-    vel_mis.w = 0;
     motor_left.control_vel = 0;
     motor_left.measure_vel = 0;
     motor_left.refer_vel = 0;
@@ -150,20 +121,7 @@ void init_parameter(void) {
     constraint.max_left = 14000;
     constraint.max_right = 14000;
 
-    const_vel[0] = 1;
-
     enable_motors = false;
-}
-
-void update_parameter_unicycle(void) {
-    parameter_unicycle_int.radius_l = ((int) (parameter_unicycle.radius_l * 1000.0));
-    parameter_unicycle_int.radius_r = ((int) (parameter_unicycle.radius_r * 1000.0));
-    parameter_unicycle_int.wheelbase = ((int) (parameter_unicycle.wheelbase * 1000.0));
-    k_odo.k_left = parameter_unicycle.radius_l * parameter_motor_left.k_ang;
-    k_odo.k_right = parameter_unicycle.radius_r * parameter_motor_right.k_ang;
-    wheel_m = parameter_unicycle.wheelbase / 2;
-    emergency.time = 1.0;
-    emergency.timeout = 500;
 }
 
 void update_parameter_motors(void) {
@@ -174,10 +132,6 @@ void update_parameter_motors(void) {
     motor_vel_right.sign = parameter_motor_right.versus;
     QEI1CONbits.SWPAB = (parameter_motor_left.versus >= 1) ? 1 : 0; // Phase A and Phase B inputs swapped
     QEI2CONbits.SWPAB = (parameter_motor_right.versus >= 1) ? 1 : 0; // Phase A and Phase B inputs swapped
-
-    //Odometry
-    k_odo.k_left = parameter_unicycle.radius_l * parameter_motor_left.k_ang;
-    k_odo.k_right = parameter_unicycle.radius_r * parameter_motor_right.k_ang;
 }
 
 void init_pid_control(void) {
@@ -227,89 +181,73 @@ void InitPid2(void) {
     PIDCoeffCalc(&kCoeffs2[0], &PIDstruct2);
 }
 
-bool Emergency(void) {
-    if (save_velocity) {
-        last_vel_rif.v = vel_rif.v;
-        last_vel_rif.w = vel_rif.w;
-        save_velocity = false;
-    }
-    vel_rif.v -= last_vel_rif.v * (((float)parameter_system.int_tm_mill) / 1000) / emergency.time;
-    vel_rif.w -= last_vel_rif.w * (((float)parameter_system.int_tm_mill) / 1000) / emergency.time;
-    if (SGN(last_vel_rif.v) * vel_rif.v < 0) vel_rif.v = 0;
-    if (SGN(last_vel_rif.w) * vel_rif.w < 0) vel_rif.w = 0;
-    if((vel_rif.v == 0) && (vel_rif.w == 0)) {
-        save_velocity = true;
-        return true;
-    }
-    return false;
-}
-
-int motorReference(motor_ref_int_t motor_ref_int) {
+int motorReference(motor_control_t motor_ref_int) {
     unsigned int t = TMR1; // Timing function
 
     MOTOR_ENABLE1 = enable_motors ^ parameter_motor_left.enable_set;
     MOTOR_ENABLE2 = enable_motors ^ parameter_motor_right.enable_set;
 
-    if (abs(motor_ref_int.ref_con_left) > constraint.max_left) {
-        motor_left.refer_vel = SGN((int)motor_ref_int.ref_con_left) * constraint.max_left;
+    if (abs(motor_ref_int.motor[0]) > constraint.max_left) {
+        motor_left.refer_vel = SGN((int) motor_ref_int.motor[0]) * constraint.max_left;
     } else {
-        motor_left.refer_vel = (int)motor_ref_int.ref_con_left;
+        motor_left.refer_vel = (int) motor_ref_int.motor[0];
     }
-    if (abs(motor_ref_int.ref_con_right) > constraint.max_right) {
-        motor_right.refer_vel = SGN((int)motor_ref_int.ref_con_right) * constraint.max_right;
+    if (abs(motor_ref_int.motor[1]) > constraint.max_right) {
+        motor_right.refer_vel = SGN((int) motor_ref_int.motor[1]) * constraint.max_right;
     } else {
-        motor_right.refer_vel = (int)motor_ref_int.ref_con_right;
+        motor_right.refer_vel = (int) motor_ref_int.motor[1];
     }
-
-    return TMR1 - t; // Time of esecution
-}
-
-motor_ref_int_t VelToMotorReference(void) {
-    motor_ref_int_t motor_ref_int;
-
-    // >>>>> Second part: references calculation
-    motor_ref_int.ref_con_left = (long int) ((1.0f / parameter_unicycle.radius_r)*(vel_rif.v + (parameter_unicycle.wheelbase * (-vel_rif.w)))*1000);
-    motor_ref_int.ref_con_right = (long int) ((1.0f / parameter_unicycle.radius_l)*(vel_rif.v - (parameter_unicycle.wheelbase * (-vel_rif.w)))*1000);
-
-    // TODO to avoid the following saturation we can normalize ref value! by Walt
-
-    // >>>>> Saturation on 16 bit values
-    motor_ref_int.ref_con_left = motor_ref_int.ref_con_left>32767?32767:motor_ref_int.ref_con_left;
-    motor_ref_int.ref_con_left = motor_ref_int.ref_con_left<-32768?-32768:motor_ref_int.ref_con_left;
-
-    motor_ref_int.ref_con_right = motor_ref_int.ref_con_right>32767?32767:motor_ref_int.ref_con_right;
-    motor_ref_int.ref_con_right = motor_ref_int.ref_con_right<-32768?-32768:motor_ref_int.ref_con_right;
-    // <<<<< Saturation on 16 bit values
-
-    return motor_ref_int;
-}
-
-int VelocityMeasure(void) {
-    unsigned int t = TMR1; // Timing function
-
-    long vel_v = (parameter_unicycle_int.radius_r * motor_right.measure_vel + parameter_unicycle_int.radius_l * motor_left.measure_vel) / 2;
-    long vel_w = (parameter_unicycle_int.radius_r * motor_right.measure_vel - parameter_unicycle_int.radius_l * motor_left.measure_vel) / (2 * parameter_unicycle_int.wheelbase);
-    vel_mis.v = ((float) vel_v / 1000000);
-    vel_mis.w = ((float) vel_w / 1000);
 
     return TMR1 - t; // Time of esecution
 }
 
 int MotorTaskController(void) {
     unsigned int t = TMR1; // Timing function
+    short i;
+    motor_control_t motor_ref_int;
 
-    /**
-     * Measure linear and angular velocity for unicycle robot
-     */
-    VelocityMeasure();
-    /**
-     * Convertion linear velocity and angular velocity to motor left and motor right
-     */
-    motor_ref_int_t motor_ref_int = VelToMotorReference();
-    /**
-     * Enable motors and check constraint and save references
-     */
-    motorReference(motor_ref_int);
+    switch (control_state) {
+        case DISABLE_HIGH_CONTROL_STATE:
+            break;
+        case VELOCITY_UNICYCLE_CONTROL_STATE:
+            /**
+             * Measure linear and angular velocity for unicycle robot
+             */
+            VelocityMeasure();
+            /**
+             * Convertion linear velocity and angular velocity to motor left and motor right
+             */
+            motor_ref_int = VelToMotorReference();
+            control_motor_state[0] = VELOCITY_CONTROL_STATE;
+            control_motor_state[1] = VELOCITY_CONTROL_STATE;
+            break;
+        case CONFIGURATION_CONTROL_STATE:
+            break;
+        default:
+            break;
+    }
+
+    for (i = 0; i < NUM_MOTORS; ++i) {
+        switch (control_motor_state[i]) {
+            case DIRECT_CONTROL_STATE:
+                //TODO To be implemet (Read issue #14)
+                break;
+            case POSITION_CONTROL_STATE:
+                //TODO to be implement
+                break;
+            case VELOCITY_CONTROL_STATE:
+                /**
+                 * Enable motors and check velocity constraint and save references
+                 */
+                motorReference(motor_ref_int);
+                break;
+            case TORQUE_CONTROL_STATE:
+                //TODO to be implement
+                break;
+            default:
+                break;
+        }
+    }
 
     return TMR1 - t; // Time of esecution
 }
@@ -333,10 +271,10 @@ int MotorPIDL(void) {
     PIDstruct1.controlReference = motor_left.refer_vel; //Riferimento Ruota Sinistra
     PIDstruct1.measuredOutput = motor_left.measure_vel; //Misura velocità
     PID(&PIDstruct1); //Esecuzione funzione PID
-    int pid_control = motor_vel_left.sign*(PIDstruct1.controlOutput >> 4) + 2049; //Conversione valore per PWM
+    int pid_control = motor_vel_left.sign * (PIDstruct1.controlOutput >> 4) + 2049; //Conversione valore per PWM
     //Invio dell'azione di controllo al motore per mezzo del PWM
     SetDCMCPWM1(1, pid_control, 0);
-    motor_left.control_vel = motor_vel_left.sign*PIDstruct1.controlOutput;
+    motor_left.control_vel = motor_vel_left.sign * PIDstruct1.controlOutput;
 
     return TMR1 - t; //Misura tempo di esecuzione
 }
@@ -360,10 +298,10 @@ int MotorPIDR(void) {
     PIDstruct2.controlReference = motor_right.refer_vel; //Riferimento Ruota Destra
     PIDstruct2.measuredOutput = motor_right.measure_vel; //Misura velocità
     PID(&PIDstruct2); //Esecuzione funzione PID
-    int pid_control = motor_vel_right.sign*(PIDstruct2.controlOutput >> 4) + 2049; //Conversione valore per PWM
+    int pid_control = motor_vel_right.sign * (PIDstruct2.controlOutput >> 4) + 2049; //Conversione valore per PWM
     //Invio dell'azione di controllo al motore per mezzo del PWM
     SetDCMCPWM1(2, pid_control, 0);
-    motor_right.control_vel = motor_vel_right.sign*PIDstruct2.controlOutput;
+    motor_right.control_vel = motor_vel_right.sign * PIDstruct2.controlOutput;
 
     return TMR1 - t; //Misura tempo di esecuzione
 }
